Finish moving assets to media

+++ mod_social_slider/script.php
@@ -254,7 +254,7 @@ class Mod_Social_SliderInstallerScript
 	}
 
 	/**
-	 * Function to update the file structure and params for the Social Slider Version 1.3.2 updates
+	 * Function to update the params for the Social Slider Version 1.3.2 updates
 	 *
 	 * @return  void
 	 *
@@ -287,7 +287,7 @@ class Mod_Social_SliderInstallerScript
 
 
 	/**
-	 * Function to update the file structure and params for the Social Slider Version 1.3.2 updates
+	 * Function to update the file structure for the Social Slider Version 1.4.0 updates
 	 *
 	 * @return  void
 	 *
@@ -299,14 +299,16 @@ class Mod_Social_SliderInstallerScript
 		jimport('joomla.filesystem.folder');
 		jimport('joomla.filesystem.file');
 
-		// Move the assets
+		// Move the assets and add index.html files to new directory
 		if (JFolder::create('media/' . $this->extension)
 			&& JFolder::move(JUri::root() . 'modules/'. $this->extension . '/assets/icons', JUri::root() . 'media/'. $this->extension)
 			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/index.html', JUri::root() . 'media/'. $this->extension . '/index.html')
 			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/jquery-sortable-min.js', JUri::root() . 'media/'. $this->extension . '/js/jquery-sortable-min.js')
 			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/jquery-sortable.js', JUri::root() . 'media/'. $this->extension . '/js/jquery-sortable.js')
 			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/jquery.js', JUri::root() . 'media/'. $this->extension . '/js/jquery.js')
-			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/style.css', JUri::root() . 'media/'. $this->extension . '/css/style.css'))
+			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/style.css', JUri::root() . 'media/'. $this->extension . '/css/style.css')
+			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/index.html', JUri::root() . 'media/'. $this->extension . '/js/index.html')
+			&& JFile::move(JUri::root() . 'modules/'. $this->extension . '/assets/index.html', JUri::root() . 'media/'. $this->extension . '/css/index.html'))
 		{
 			// We can now delete the folder
 			JFolder::delete(JPATH_ROOT . '/modules/'. $this->extension . '/assets');
